OneDrive/Documents/Projects/Travel Bot/requirements.txt
requests
streamlit
python-dotenv
google-genai
